Large Database Update
Added institution table that stores consumerKey, shared secret, admin ID
and timezone for each institution. Currently using a default institution
with consumerKey 'Institution.A'. UserID 11112222 is the SuperAdmin used
in the exampledata.
CourseID key replaced by [CourseID,ConsumerKey], updated queries.php to
reflect this. Currently, Consumer key is set to default 'Institution.A'
in most queries.
Increased size of userID to allow for more unique usernames. All users
are now associated with an institution, but this is not an identifying
relationship as it is with courses.